finish

+++ src/models/Schedule.js
@@ -1,29 +1,18 @@
 const db = require('../utils/db')
+
 module.exports = {
-  CreateBus: function (idAgent, name, busSeat) {
-    const table = 'busses'
-    const query = `INSERT INTO ${table} (id_agent,car_name,bus_seat) VALUES (${idAgent}, '${name}', ${busSeat})`
+  createSchedule: function (idUser, idBus, idRoute, price, departureTime, arriveTime, departureDate) {
+    const table = 'schedules'
+    const query = `INSERT INTO ${table} (id_admin, id_bus, id_route, price, departure_time, arrive_time, departure_date)
+    VALUES (${idUser}, ${idBus},${idRoute}, ${price},'${departureTime}', '${arriveTime}', '${departureDate}')`
+    console.log(query)
     return new Promise(function (resolve, reject) {
       db.query(query, function (err, results, fields) {
         if (err) {
           reject(err)
         } else {
-          resolve(results.insertId)
-        }
-      })
-    })
-  },
-  updateBuss: function (id, name, size) {
-    return new Promise(function (resolve, reject) {
-      const table = 'busses'
-      const query = `UPDATE ${table} SET car_name='${name}', bus_seat=${size} WHERE id=${id}`
-      console.log(query)
-      db.query(query, function (err, results, fields) {
-        if (err) {
-          reject(err)
-        } else {
-          if (results.affectedRows) {
-            resolve(results)
+          if (results.insertId) {
+            resolve(true)
           } else {
             resolve(false)
           }
@@ -31,28 +20,24 @@ module.exports = {
       })
     })
   },
-  updateBussAdmin: function (id, idAgent, name, size) {
+  getScheduleByIdAdmin: function (idAdmin) {
+    const table = 'schedules'
+    const query = `SELECT * FROM ${table} WHERE id_admin = ${idAdmin}`
+    console.log(query)
     return new Promise(function (resolve, reject) {
-      const table = 'busses'
-      const query = `UPDATE ${table} SET id_agent=${idAgent}, car_name='${name}', bus_seat=${size} WHERE id=${id}`
-      console.log(query)
       db.query(query, function (err, results, fields) {
         if (err) {
           reject(err)
         } else {
-          if (results.affectedRows) {
-            resolve(results)
-          } else {
-            resolve(false)
-          }
+          resolve(results)
         }
       })
     })
   },
-  deleteBuss: function (id) {
+  deleteSchedule: function (id, idAdmin) {
     return new Promise(function (resolve, reject) {
-      const table = 'busses'
-      const query = `DELETE FROM ${table} WHERE id=${id}`
+      const table = 'schedules'
+      const query = `DELETE FROM ${table} WHERE id=${id} AND id_admin=${idAdmin}`
       db.query(query, function (err, results, fields) {
         if (err) {
           reject(err)
@@ -66,13 +51,13 @@ module.exports = {
       })
     })
   },
-  getAllBusses: function (conditions = {}) {
+  getAllSchedules: function (conditions = {}) {
     let { page, perPage, sort, search } = conditions
     page = page || 1
     perPage = perPage || 5
-    sort = sort || { key: 'id', value: 0 } // value => 0 untuk ascending, 1 descending
-    search = search || { key: 'car_name', value: '' }
-    const table = 'busses'
+    sort = sort || { key: 'id', value: 1 } // value => 0 untuk descending, 1 ascending
+    search = search || { key: 'name', value: '' }
+    const table = 'schedules'
     return new Promise(function (resolve, reject) {
       const sql = `SELECT * FROM ${table}
                   WHERE ${search.key} LIKE '${search.value}%'
@@ -87,10 +72,10 @@ module.exports = {
       })
     })
   },
-  getTotalBusses: function (conditions = {}) {
+  getTotalSchedules: function (conditions = {}) {
     let { search } = conditions
-    search = search || { key: 'car_name', value: '' }
-    const table = 'busses'
+    search = search || { key: 'name', value: '' }
+    const table = 'schedules'
     return new Promise(function (resolve, reject) {
       const sql = `SELECT COUNT (*) AS total FROM ${table}
                   WHERE ${search.key} LIKE '${search.value}%'`
@@ -103,31 +88,18 @@ module.exports = {
       })
     })
   },
-  findBusByAgent: function (id) {
-    const table = 'busses'
-    const query = `SELECT * FROM ${table} WHERE id_agent=${id}`
-    console.log(query)
+  updateSchedule: function (id, idAdmin, idBus, idRoute, departureTime, arriveTime, departureDate) {
     return new Promise(function (resolve, reject) {
+      const table = 'schedules'
+      const query = `UPDATE ${table} SET id_bus=${idBus}, id_route=${idRoute},
+      departure_time='${departureTime}', arrive_time='${arriveTime}', departure_date='${departureDate}' WHERE id=${id} AND id_admin = ${idAdmin}`
+      console.log(query)
       db.query(query, function (err, results, fields) {
         if (err) {
           reject(err)
         } else {
-          resolve(results)
-        }
-      })
-    })
-  },
-  findBusById: function (id) {
-    const table = 'busses'
-    const query = `SELECT * FROM ${table} WHERE id=${id}`
-    console.log(query)
-    return new Promise(function (resolve, reject) {
-      db.query(query, function (err, results, fields) {
-        if (err) {
-          reject(err)
-        } else {
-          if (results.length) {
-            resolve(results[0])
+          if (results.affectedRows) {
+            resolve(true)
           } else {
             resolve(false)
           }
@@ -135,17 +107,34 @@ module.exports = {
       })
     })
   },
-  updateBussSeat: function (id) {
+  getScheduleById: async function (id) {
+    const table = 'schedules'
+    const query = `SELECT * FROM ${table} WHERE id = ${id}`
+    console.log(query)
     return new Promise(function (resolve, reject) {
-      const table = 'busses'
-      const query = `UPDATE ${table} SET bus_seat=bus_seat-1 WHERE id=${id}`
-      // console.log(query)
       db.query(query, function (err, results, fields) {
         if (err) {
           reject(err)
         } else {
-          if (results.affectedRows) {
-            resolve(results)
+          if (results.length) {
+            resolve(true)
+          } else {
+            resolve(false)
+          }
+        }
+      })
+    })
+  },
+  getPriceBySchedule: async function (idSchedule) {
+    const table = 'schedules'
+    const query = `SELECT price FROM ${table} WHERE id = ${idSchedule}`
+    return new Promise(function (resolve, reject) {
+      db.query(query, function (err, results, fields) {
+        if (err) {
+          reject(err)
+        } else {
+          if (results.length) {
+            resolve(results[0])
           } else {
             resolve(false)
           }
